Pandemic Tweet Challenge Doc
Social media has become the main source of communication and a method of extracting public opinions and certain topics related to the nature of the dataset. The problem we are trying to solve is deriving meaningful insight into public opinion regarding largely controversial issues. This dataset is regarding COVID-19 and is a controversial topic on different levels whether lockdowns are placed in action, vaccines, or mask mandates. It is difficult to analyze these opinions on a surface level. Building a model that gives insight into the source of controversy in these public opinions gives public officials the ability to speak to them with greater understanding on the topic, demographics, segments, and trends.
Sentiment analysis is a widely used method currently used to find accuracy of the application being used. Although widely used it faces challenges such as polarity, sarcasm, comparative sentences, emojis; our goal is to create a model which can learn from these issues and improve as the model is trained. We plan to use the LSTM model and Fully Connected Feedforward model because both allow for a training component to sentiment analysis for the different tweets. Both models implement the technique known as gradient descent over random forest and gradient boosting leading to successful predictions and higher accuracies.